feat: Update piquasso to version 2.0.0
