+++ piquassoboost/fock/general/simulator.py
@@ -15,8 +15,8 @@
 
 import piquasso as pq
 
-from piquassoboost.config import BoostConfig
+from piquassoboost.calculator import BoostCalculator
 
 
 class BoostedFockSimulator(pq.FockSimulator):
-    _config_class = BoostConfig
+    _calculator_class = BoostCalculator
