+++ piquassoboost/sampling/simulator.py
@@ -18,6 +18,7 @@ import piquasso as pq
 from piquasso.instructions import measurements
 
 from piquassoboost.config import BoostConfig
+from piquassoboost.calculator import BoostCalculator
 
 from .calculations import particle_number_measurement
 
@@ -29,3 +30,4 @@ class BoostedSamplingSimulator(pq.SamplingSimulator):
     }
 
     _config_class = BoostConfig
+    _calculator_class = BoostCalculator
